+++ src/circuit/states/super_positions.rs
@@ -61,19 +61,18 @@ impl SuperPosition {
     ///
     /// assert_eq!(&complex_re_array![1f64, 0f64, 0f64, 0f64], superpos.get_amplitudes());
     /// ```
-    pub fn new_with_amplitudes(amplitudes: &[Complex<f64>]) -> QResult<SuperPosition> {
+    pub fn new_with_amplitudes(amplitudes: &[Complex<f64>]) -> QResultConst<SuperPosition> {
         if !Self::equal_within_error(amplitudes.iter().map(|x| x.abs_square()).sum::<f64>(), 1f64) {
-            return Err(QuantrError {
-                message: String::from("Slice given to set amplitudes in super position does not conserve probability, the absolute square sum of the coefficents must be one."),
+            return Err(QuantrErrorConst {
+                message: "Slice given to set amplitudes in super position does not conserve probability, the absolute square sum of the coefficents must be one.",
             });
         }
 
         let length = amplitudes.len();
         if (length & (length - 1)) != 0 {
-            return Err(QuantrError {
-                message: String::from(
+            return Err(QuantrErrorConst {
+                message: 
                     "The length of the array must be of the form 2**n where n is an integer.",
-                ),
             });
         }
 
@@ -126,7 +125,8 @@ impl SuperPosition {
         })
     }
 
-    /// Retrieves the coefficient of the product state in the computational basis given by the list index.
+    /// Retrieves the coefficient of the product state in the computational basis given by the list index. Returns `None` if the
+    /// index is greater than the product dimension of the superposition.
     ///
     /// # Example
     /// ```
@@ -137,14 +137,8 @@ impl SuperPosition {
     ///
     /// assert_eq!(complex_re!(1f64), superpos.get_amplitude(1).unwrap());
     /// ```
-    pub fn get_amplitude(&self, pos: usize) -> QResult<Complex<f64>> {
-        if pos >= self.amplitudes.len() {
-            let length = self.amplitudes.len();
-            Err(QuantrError { message: format!("Failed to retrieve amplitude from list. Index given was, {pos}, which is greater than length of list, {length}."), 
-            })
-        } else {
-            Ok(*self.amplitudes.get(pos).unwrap())
-        }
+    pub fn get_amplitude(&self, pos: usize) -> Option<Complex<f64>> {
+        self.amplitudes.get(pos).cloned()
     }
 
     /// Returns the number of qubits that each product state in the super position is composed of by using the Kronecker product.
@@ -158,7 +152,7 @@ impl SuperPosition {
     ///
     /// assert_eq!(2, superpos.get_num_qubits());
     /// ```
-    pub fn get_num_qubits(&self) -> usize {
+    pub const fn get_num_qubits(&self) -> usize {
         self.product_dim
     }
 
@@ -209,7 +203,7 @@ impl SuperPosition {
         if 2usize << (prod_state.qubits.len() - 1) != self.amplitudes.len() {
             return Err(QuantrError { message: format!("Unable to retreive product state, |{:?}> with dimension {}. The superposition is a linear combination of states with different dimension. These dimensions should be equal.", prod_state.to_string(), prod_state.num_qubits()),});
         }
-        Ok(*self.amplitudes.get(prod_state.comp_basis()).unwrap())
+        Ok(self.amplitudes[prod_state.comp_basis()]) 
     }
 
     /// Returns a new superposition in the computational basis.